Reorganise routing structure
