+++ app/(auth)/welcome/_components/Hero.tsx
