+++ app/(auth)/welcome/layout.tsx
