+++ app/(auth)/welcome/page.tsx
@@ -1,5 +1,5 @@
-import Main from "../_components/wrappers/PageMain";
-import RowWrapper from "../_components/wrappers/RowWrapper";
+import Main from "../../_components/wrappers/PageMain";
+import RowWrapper from "../../_components/wrappers/RowWrapper";
 import Hero from "./_components/Hero";
 import Introduction from "./_components/Introduction";
 import LandingForm from "./_components/LandingForm";
